google-assistant-webserver/requirements.txt
aiohttp[speedups]
google-assistant-sdk[samples]
google-auth-oauthlib[tool]
uvloop